+++ src/lib/meson.build
@@ -15,10 +15,10 @@ source_file = [
    'tvgBinaryDesc.h',
    'tvgFill.h',
    'tvgLoader.h',
-   'tvgLoaderMgr.h',
+   'tvgLoadModule.h',
    'tvgPictureImpl.h',
    'tvgRender.h',
-   'tvgSaver.h',
+   'tvgSaveModule.h',
    'tvgSceneImpl.h',
    'tvgShapeImpl.h',
    'tvgTaskScheduler.h',
@@ -28,7 +28,7 @@ source_file = [
    'tvgGlCanvas.cpp',
    'tvgInitializer.cpp',
    'tvgLinearGradient.cpp',
-   'tvgLoaderMgr.cpp',
+   'tvgLoader.cpp',
    'tvgPaint.cpp',
    'tvgPicture.cpp',
    'tvgRadialGradient.cpp',
