+++ src/lib/tvgInitializer.cpp
@@ -21,7 +21,7 @@
  */
 #include "tvgCommon.h"
 #include "tvgTaskScheduler.h"
-#include "tvgLoaderMgr.h"
+#include "tvgLoader.h"
 
 #ifdef THORVG_SW_RASTER_SUPPORT
     #include "tvgSwRenderer.h"
