+++ src/lib/tvgLoadModule.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Samsung Electronics Co., Ltd. All rights reserved.
+ * Copyright (c) 2020-2021 Samsung Electronics Co., Ltd. All rights reserved.
 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,29 +19,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#ifndef _TVG_SAVER_H_
-#define _TVG_SAVER_H_
+#ifndef _TVG_LOAD_MODULE_H_
+#define _TVG_LOAD_MODULE_H_
 
-#include "tvgPaint.h"
+#include "tvgCommon.h"
 
 namespace tvg
 {
 
-class SaveModule
+class LoadModule
 {
 public:
-    virtual ~SaveModule() {}
+    //default view box, if any.
+    float vx = 0;
+    float vy = 0;
+    float vw = 0;
+    float vh = 0;
+    float w = 0, h = 0;         //default image size
+    bool preserveAspect = true; //keep aspect ratio by default.
 
-    virtual bool save(Paint* paint, const string& path) = 0;
+    virtual ~LoadModule() {}
+
+    virtual bool open(const string& path) { /* Not supported */ return false; };
+    virtual bool open(const char* data, uint32_t size, bool copy) { /* Not supported */ return false; };
+    virtual bool open(const uint32_t* data, uint32_t w, uint32_t h, bool copy) { /* Not supported */ return false; };
+    virtual bool read() = 0;
     virtual bool close() = 0;
-
-    //Utility Method: Iterator Delegator
-    Iterator* iterator(const Paint* paint)
-    {
-        return paint->pImpl->iterator();
-    }
+    virtual const uint32_t* pixels() { return nullptr; };
+    virtual unique_ptr<Scene> scene() { return nullptr; };
 };
 
 }
 
-#endif //_TVG_SAVER_H_
+#endif //_TVG_LOAD_MODULE_H_
