+++ src/loaders/jpg/tvgJpgLoader.h
@@ -25,13 +25,13 @@
 using tjhandle = void*;
 
 //TODO: Use Task?
-class JpgLoader : public Loader
+class JpgLoader : public LoadModule
 {
 public:
     JpgLoader();
     ~JpgLoader();
 
-    using Loader::open;
+    using LoadModule::open;
     bool open(const string& path) override;
     bool open(const char* data, uint32_t size, bool copy) override;
     bool read() override;
