+++ src/savers/tvg/tvgTvgSaver.cpp
@@ -22,7 +22,7 @@
 #include <float.h>
 #include <math.h>
 #include <stdio.h>
-#include "tvgSaver.h"
+#include "tvgSaveModule.h"
 #include "tvgTvgSaver.h"
 
 #define SIZE(A) sizeof(A)
